Created mapper module to separate from calculator class
